examples/nims/genmol/requirements.txt
pandas
numpy
rdkit
safe-mol
datamol
matplotlib
